Sign-up fix

+++ frontend/src/APIClients/BaseAPIClient.ts
@@ -3,6 +3,8 @@ import {
   ApolloLink,
   InMemoryCache,
   Observable,
+  ObservableSubscription,
+  Operation,
 } from "@apollo/client";
 import { onError } from "@apollo/client/link/error";
 import { setContext } from "@apollo/client/link/context";
@@ -12,8 +14,6 @@ import { createUploadLink } from "apollo-upload-client";
 import jwt from "jsonwebtoken";
 import AUTHENTICATED_USER_KEY from "../constants/AuthConstants";
 import { getLocalStorageObjProperty } from "../utils/LocalStorageUtils";
-
-/* eslint-disable @typescript-eslint/no-explicit-any */
 
 const httpLink = (createUploadLink({
   uri: `${process.env.REACT_APP_BACKEND_URL}/graphql`,
@@ -32,7 +32,8 @@ const authFromLocalLink = setContext(async (_, { headers }) => {
     },
   };
 });
-const injectAccessToken = async (operation: any) => {
+
+const injectAccessToken = async (operation: Operation) => {
   const accessToken = getLocalStorageObjProperty(
     AUTHENTICATED_USER_KEY,
     "accessToken",
@@ -40,12 +41,15 @@ const injectAccessToken = async (operation: any) => {
   operation.setContext({
     headers: { Authorization: `Bearer ${accessToken}` },
   });
-  const decodedToken: any = jwt.decode(String(accessToken));
+  const decodedToken = jwt.decode(String(accessToken));
   const currTime = Math.round(new Date().getTime() / 1000);
-  if (!decodedToken || decodedToken.exp <= currTime) {
+  if (typeof(decodedToken) === 'object' && decodedToken?.exp <= currTime) {
     localStorage.clear();
     window.location.reload();
     throw Error("Token expired, sign out");
+  }
+  if (!decodedToken) {
+    console.warn('null access token - failed to inject into API call');
   }
 };
 
@@ -53,7 +57,7 @@ const injectAccessToken = async (operation: any) => {
 const accessTokenInjectionLink = new ApolloLink(
   (operation, forward) =>
     new Observable((observer) => {
-      let handle: any;
+      let handle: ObservableSubscription | null;
       Promise.resolve(operation)
         .then((op) => injectAccessToken(op))
         .then(() => {
@@ -64,15 +68,13 @@ const accessTokenInjectionLink = new ApolloLink(
           });
         })
         .catch(observer.error.bind(observer));
-      return () => {
-        if (handle) handle.unsubscribe();
-      };
+      return () => handle?.unsubscribe();
     }),
 );
 
 const refreshDirectionalLink = new RetryLink().split(
   (operation) =>
-    ["Refresh", "ResetPassword", "Login", "SignUp", "LoginWithGoogle"].includes(
+    ["Refresh", "ResetPassword", "Login", "SignUp"].includes(
       operation.operationName,
     ),
   authFromLocalLink.concat(httpLink),
